refactor: extract menu events from App
